requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
boto3==1.42.9
botocore==1.42.9
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
fastapi==0.124.4
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.8
multidict==6.7.0
multiprocess==0.70.16
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
